Snapshot of work-in-progress on collapsing menus.
I _think_ the navigator class has great potential but it must behave itself... I'm fairly frustrated I wasted many hours yesterday on an overly complicated approach; I've removed the notion of inner/outer/select context from the BFS visitor algorithm (which is now actually BFS BTW) and replaced it with an optional, externally-specified, context structure that is passed back to the _single_ callback. The visitor callback can scribble whatever it wants in the context object it provides for itself. Currently I am not leveraging this facility. Other changes: I'm replace the topLevelMenuLevels array with a root instance of NavigatorWindowMenuLevel constructed with a modified layout. This simplifies everything considerably. The outer call stack is hooked up and looks okay at this point. I'm about to try to implement the actual visitor algorithm that's responsible for deciding if a specific menu level is visible or not based on: the currently-selected menu level object, and the inner/outer detail observable values.