Navigator now allows multiple extension objects to be added to an extension point array.
Previously, multiple adds issued from an "extension"-type item host object would simply splice in
references to the same object (almost what I have in mind for the clone operation). Now, if the
"select"-type item host child of the "extension"-type item host parent is in the "element"
state, the item is detached from the parent extension array, and re-initialized prior to add
so that we get distinct objects stored in the extension array.