Navigator member namespace edit is working okay.
Note that there's some refactoring (update on a menu item host should call a declared callback to affect change of model view members (the navigator lib should be decoupled from whatever the hell the schema of the object model it navigates is). Also, edit is currently only via a textarea which works okay for strings but is inappropriate for other more complex types (e.g. uuidSelection).