Snapshot on work to refactor navigator into modular pieces.
In more detail, I'm working on a so-called OMM (Object Model Manager) that abstracts an in-memory Javascript object store. OMM encapsulates (a) object model schema (deduced from object model declaration) (b) the in-memory store. The intent is to provide a low-overhead addressable object store that maintains pure user data (and a small amount of extra state required to track changes on behalf of _clients_ that react to changes.