Work on OMM core.
This commit contains work on class ObjectModel that traverses a hierarchy of declared object descriptors contained within an object model "layout" declaration to deduce several contained data structures. Currently the traversal leverages recursion and produces a map of object model path to path descritor value and an orderred array of the paths.