Snapshot of work on OMM ObjectModel.
ObjectModel's generated object model descriptors now include an orderred array of path ID's that correspond to the extension points that must be resolved in order to actually address anything using an object model path. For example:

Suppose a.b.c where a is the root, b is an extension point, c is an archetype (as declared in the layout). c is an array element of b's array. in order to address something at a.b.c, an index (or key that can be resolved to an index) must be used to select a specific c from collection b. We would know this by looking at c.pathResolveExtensionPoints, seeing the array has a single element, obtaining it's value (e.g. 1 that corresponds to path a.b. A function resolving a.b.c would thus require a vector containing one index (or key) in order to return a reference to whatever a.b.c stands for in object model path space.