Continue double-down on concept of object model manager.
This commit includes work primarily on the ObjectStoreNamespaceBinder class. This implementation leverages the pre-computed parent path cached in the path ID's descriptor (obtained from ObjectModel) to implement a "bind" operation that returns an instance of ObjectStoreNamespace if it's able to successfully resolve the ObjectModelNamespaceSelector coordinates within the ObjectStore. Note that bind will always succeed for ObjectModelNamespaceSelector objects that require no extension keys (i.e. the object model namespace selector selects a namespace in the root component of the object model). Currently, the binder will create object namespaces that do not already exist within the ObjectStore provided that ObjectModelNamespaceSelector can be resolved as above.