Snapshot of ObjectStoreNamespace.internalResolveNamespaceDescriptor work.
ObjectStoreNamespace.internalResolveNamespaceDescriptor method now requires the mode_ input parameter be one of "bypass" | "new" | "strict". The action of this method, obtaining a reference to a specific sub-object (aka "store namespace"), is predicated on the specified mode. "new" creates missing namespaces in the store as required to resolve the object model namespace selector whereas "bypass" and "strict" presume the prior existence of selected namespace(s). In the case of "strict", namespace members are verified.