ObjectStore.removeComponent implementation.
This commit provides implementation of ObjectStore methods createComponent, removeComponent, and openNamespace and toJSON. Note that ObjectStoreNamespace has been modified to track ResolvedKeyIndexVector. Currently this information is leveraged to determine the index of a component that is being removed. However, this information can additionally be used to reduce overhead of resolving store namespaces to constant (from linear currently).