Change window manager view model to hide instead of removing inactive layout planes.
This change is intended to address a problem discovered during prelimary experiments with D3js that remove the rendered SVG on a window manager library view plane switch. Previously the model view template leveraged Knockout.js if binding (removing inactive planes from the DOM entirely) and now is simply leverages the Knockout.js visible binding (leaves the rendered content in the DOM but hides it). This eliminates the need to track plane switches and improves rendering performance at the cost of some DOM memory overhead (which I'm not overly concerned about at this point anyway).