ObjectStore/observer callback testing for component add/remove.
This commit implements broadcast callbacks to model view objects registered with an ObjectStore instance as an observer on creation of a component as well as removal. Removal of a leaf component appears to work in preliminary tests but am still working to ensure removal of parent component correctly unreifies all removed sub-components (if there are any).