Add ObjectModelNavigatorSelectorWindow class, integrate with SelectorStore notifications.
This commit contains a simple Knockout.js MVVM presentation class ObjectModelModelNavigatorSelectorWindow that leverages SelectorStore onComponentUpdate notifications to update a clickable selector path view. Note that this model view is not the actual selector store but rather an observer of the selector store. Clicking on path component doesn't direcly modify the state of the ObjectModelNavigatorSelectorWindow. Rather, it makes a call on the ObjectStore to set a new selector and the SelectorStore calls the ObjectModelNavigatorSelectorWindow back which triggers the presentation state update to display the current selection path.