More work on adding recursive objects in OMM.
This commit captures work on the add component path. A large part of the call path is now converted over to understand the additions made to namespace selectors that allow them to address into recursive object hierarchies. Next up is to complex the add path by updating the observer routines to understand how to traverse around inside these structures. Then likely a bunch of work on the model views to get it all to _really_ work correctly.