More work on recursively declared components in OM
This commit focusses on getting the object model navigator model view to work correctly (involves ensuring that all the traversals that occur over a store that result in observer notifications are working and that the observers also do the right thing with the selectors they're passed). This version adds a recursive component (looks okay in the JSON serialization) and gets the menus correctly setup in the navigator host.