Work on helper functions for creating omm.Address instances.
Everything is still disconnected. However, the new omm.Namespace is working essentially (some support methods need to be added back and ported to the new addressing scheme). The current work on manipulating omm.Address objects to make the porting nice and clean. Most operations involved in traversal of the store namespace hierarchy by omm.Store involve the manipulation of addresses. Past iterations of this library have left a lot of detail exposed.