Snapshot of work in progress on ONMjs.Address...
... I've started reworking the reification routines (i.e. the routines that "make the store real" to observers by calling their registered observer interfaces) and upon review see that the majority of this code can be refactored.

Firstly, the actual dispatch of callbacks is little tricky due there being two distinct behaviors: single observer dispatch or broadcast to all registered observers. This logic has been moved off into a subroutine of ONMjs.implementation.StoreReifier. This leaves in the body of the reification methods primary code that, in the old regime, performed rather trivial operations on "selectors" (what ONMjs.Address replaces) in order to affect traversal of the store structure. Okay. Well now the store structure is a whole lot more complicated potentially with the introduction of Address and its support for recursively extensible component hierarchies. So...

Traveral ain't like it used to be. But, it must be simple and efficient of the implentors of observer routines to use the Address they're passed. And, operations such as those performed by Store's reifier aren't uncommon. So traveral logic is being built into ONMjs.Address and exposed using the visitor pattern.

More on this later.