Okay... we're turning the corning on this reconstruction effort...
This commit features the ONMjs.observers.NavigatorModelView working end-to-end eith ONJS.Store and ONMjs.CachedAddress via observer registrations. Clicking on one of the contained ONMjs.observer.NavigatorItemModelView items routes a call to ONMjs.CachedAddress which then broadcasts an update that the Navigator responds to by modifying the state of the contained Item. So the act of simply clicking in the Navigator and having the selection track demonstrates a huge portion of the revamped ONMjs library working. I am bound to find some more bugs and further refine the API as I port the remaining observers over to the new addressing protocol (and what in the end is a near total re-write of the library... that's okay - it's much much cleaner now).