Solid progress on address fixes...
... Today I've rewritten AddressTokenBinder to improve efficiency and fix bug in assignment of the key to the resolved token.

Performance enhance ONMjs.Namespace (don't maintain a separate array of data references - it's not used and we have the information if we need it in the token binder array.

Currently hunting what I think is the last bug: when an unqualified component address with more than one unqualified tail tokens is passed to ONMjs.Store.createComponent we're not yet handling all the possible create component and extension point namespace update events correctly. Note that this is fine if you create components one level at at time. But breaks observers if the events are not dispatched correctly.